Use sin-phase kernels for Humonica resynthesis
Instead of setting all phases to 0 (cosine) which gives a central peak,
and even symmetry around 0, we now set the phases to pi/2 (sine).

This sounds the same, but it eliminates any DC-offset issues, because
the symmetry is now odd, so any symmetric windowing of the kernel will
have zero DC offset.